+++ src/Charts/Donut/styles.scss
