+++ src/Charts/Gauge/styles.scss
