+++ src/PresentationalComponents/Ansible/ansible.js
@@ -3,7 +3,7 @@ import propTypes from 'prop-types';
 
 import classNames from 'classnames';
 
-import './ansible.scss';
+import './styles.scss';
 
 /**
  * This is a dumb component that only recieves properties from a smart component.
