+++ src/PresentationalComponents/Ansible/index.js
@@ -1,1 +1,1 @@
-export { default as Ansible } from './ansible';
+export { default as Ansible } from './Ansible';
