+++ src/PresentationalComponents/PageHeader/index.js
@@ -1,2 +1,2 @@
-export {default as PageHeaderTitle } from './page-header-title';
-export {default as PageHeader } from './page-header';
+export { default as PageHeaderTitle } from './PageHeaderTitle';
+export { default as PageHeader } from './PageHeader';
