+++ src/PresentationalComponents/SampleComponent/styles.scss
