+++ src/PresentationalComponents/Section/index.js
@@ -1,1 +1,1 @@
-export { default as Section } from './section.js';
+export { default as Section } from './Section.js';
