+++ src/index.js
@@ -1,9 +1,6 @@
-
-export { default as PageHeader } from './PresentationalComponents/PageHeader/page-header.js';
-export { default as PageHeaderTitle } from './PresentationalComponents/PageHeader/page-header-title.js';
-export { default as Section } from './PresentationalComponents/Section/section.js';
-export { default as SampleComponent } from './PresentationalComponents/SampleComponent/sample-component.js';
-export { default as Ansible } from './PresentationalComponents/Ansible/ansible.js';
+export { default as Section } from './PresentationalComponents/Section/Section.js';
+export { default as SampleComponent } from './PresentationalComponents/SampleComponent/SampleComponent.js';
+export { default as Ansible } from './PresentationalComponents/Ansible/Ansible.js';
 export { default as Main } from './PresentationalComponents/Main/Main.js';
 export {
     default as ReducerRegistry,
@@ -13,6 +10,7 @@ export {
 } from './Utilities/ReducerRegistry';
 export { default as MiddlewareListener } from './Utilities/MiddlewareListener';
 export { default as registry, getStoreFromRegistry } from './Utilities/Registry';
+export * from './PresentationalComponents/PageHeader';
 export * from './PresentationalComponents/Pagination';
 export * from './PresentationalComponents/SimpleTableFilter';
 export * from './PresentationalComponents/Input';
